Add EqualizerConfiguration and data saving
The saving and loading should already be wired into EqualizerViewController, hopefully correctly. Also fiddled with some more things, added some more API-type functions, and left more notes and marks. Take a look at the diff versus the previous commit for more specifics. Running the app on my end, it is functional at the same level as the demo version (which is good - nothing visibly broken during heavy refactoring)

+++ equal sounds/EqualizerBand.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum EqualizerBand: Int, CaseIterable
+enum EqualizerBand: Float, CaseIterable
 {
     case thirtyOneHz = 31
     case sixtyTwoHz = 62
@@ -23,6 +23,6 @@ enum EqualizerBand: Int, CaseIterable
 
 extension EqualizerBand
 {
-    static let baseBand: Int = EqualizerBand.allCases.first?.rawValue ?? 0
+	static let baseBand: Float = EqualizerBand.allCases.first?.rawValue ?? 0
 }
